cache-service/requirements.txt
# cache-service/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
redis==5.0.1
numpy==1.25.2
httpx==0.25.0
pydantic==2.4.2
opentelemetry-api==1.21.0
opentelemetry-sdk==1.21.0
opentelemetry-instrumentation-fastapi==0.42b0
opentelemetry-exporter-otlp-proto-grpc==1.21.0